+++ src/lib.ts
@@ -31,8 +31,8 @@ export interface ErrorSignal {
   code: string
 }
 
-export type DescriptionSignal = RTCSessionDescriptionInit
-export type CandidateSignal = RTCIceCandidateInit | undefined
+export type DescriptionSignal = RTCSessionDescription
+export type CandidateSignal = RTCIceCandidate | undefined
 
 export function pause(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms))